Corrected error in DataWriter.saveResultsTable

+++ src/Math/Optimisation/PlateFitter.java
@@ -20,7 +20,7 @@ public class PlateFitter extends Fitter {
     private Plate plateTemplate;
     private ImageProcessor image;
 
-    public PlateFitter(ImageProcessor image, int rows, int cols, int wellRad, double xBuff, double yBuff, double interWellSpacing, double shrinkFactor) {
+    public PlateFitter(ImageProcessor image, int rows, int cols, double wellRad, double xBuff, double yBuff, double interWellSpacing, double shrinkFactor) {
         super();
         this.image = image;
         this.plateTemplate = new Plate(rows, cols, wellRad, xBuff, yBuff, interWellSpacing, shrinkFactor);
